start outfit functionality
