+++ client/src/components/RelatedProducts/Cards/Comparison.jsx
@@ -1,8 +1,6 @@
-import { React, useContext } from 'react';
+import { React } from 'react';
 import styled from 'styled-components';
-import sampleFeatures from '../sampledata/sampleFeatures';
-import Features from './Features';
-import { ProductIDContext } from '../../../contexts/ProductIDContext';
+// import sampleFeatures from '../sampledata/sampleFeatures';
 import { useFeature } from './FeatureContext';
 
 // should get an array with two arrays of features from each item
@@ -16,18 +14,37 @@ function Comparison() {
     const item1 = twoProductsArray.data[0].data;
     const item2 = twoProductsArray.data[1].data;
 
-    console.log(item1, 'this is item1');
-    console.log(item2, 'this is item2');
+    const FeatArray = item1.features.concat(item2.features);
+
+    const item1Array = [];
+    item1.features.map((item) => item1Array.push(item));
+
+    const item2Array = [];
+    item2.features.map((item) => item2Array.push(item));
 
     return (
-      <Testing>
-        <div>
+      <table>
+        <thead>
+          <tr>
+            <th>{item1.name}</th>
+            <th> </th>
+            <th>{item2.name}</th>
+          </tr>
+        </thead>
+        <tbody>
+          {FeatArray.map((item, key) => (
+            <tr key={key}>
 
-          <Features item1={item1} item2={item2} />
+              <td>{item1Array.map((item1) => (item1.feature === item.feature ? 'X' : ' '))}</td>
 
-        </div>
-      </Testing>
+              <td>{item.value ? `${item.feature}: ${item.value}` : `${item.feature}`}</td>
 
+              <td>{item2Array.map((item2) => (item2.feature === item.feature ? 'X' : ' '))}</td>
+            </tr>
+          ))}
+
+        </tbody>
+      </table>
     );
   }
 }
