+++ client/src/components/RelatedProducts/Outfits/OutfitCard.jsx
@@ -1,0 +1,10 @@
+import { React, useEffect } from 'react';
+import styled from 'styled-components';
+
+function OutfitCard() {
+  return (
+    <div>this is just a test</div>
+  );
+}
+
+export default OutfitCard;
